+++ packages/kbn-optimizer/limits.yml
@@ -98,7 +98,7 @@ pageLoadAssetSize:
   lens: 57135
   licenseManagement: 41817
   licensing: 29004
-  links: 44490
+  links: 8000
   lists: 22900
   logsDataAccess: 16759
   logsExplorer: 60000
